Add Main Menu and css styles
